chore!: changed easing function types from double to float

+++ Plazma/EasingFunctions.cs
@@ -17,26 +17,26 @@ public static class EasingFunctions
     /// <param name="c">The amount of change. (end - start).</param>
     /// <param name="d">The total amount of time(milliseconds)/steps.</param>
     /// <returns>The result of the easing function.</returns>
-    public static double EaseOutBounce(double t, double b, double c, double d)
+    public static float EaseOutBounce(float t, float b, float c, float d)
     {
-        t /= 1000.0; // Convert to seconds
-        d /= 1000.0; // Convert to seconds
+        t /= 1000.0f; // Convert to seconds
+        d /= 1000.0f; // Convert to seconds
 
         if ((t /= d) < 0.36363636363636363636363636363636)
         {
-            return (c * (7.5625 * t * t)) + b;
+            return (c * (7.5625f * t * t)) + b;
         }
         else if (t < 0.72727272727272727272727272727273)
         {
-            return (c * ((7.5625 * (t -= 0.54545454545454545454545454545455) * t) + 0.75)) + b;
+            return (c * ((7.5625f * (t -= 0.54545454545454545454545454545455f) * t) + 0.75f)) + b;
         }
         else if (t < 0.90909090909090909090909090909091)
         {
-            return (c * ((7.5625 * (t -= 0.81818181818181818181818181818182) * t) + 0.9375)) + b;
+            return (c * ((7.5625f * (t -= 0.81818181818181818181818181818182f) * t) + 0.9375f)) + b;
         }
         else
         {
-            return (c * ((7.5625 * (t -= 0.9) * t) + 0.95454545454545454545454545454545)) + b;
+            return (c * ((7.5625f * (t -= 0.9f) * t) + 0.95454545454545454545454545454545f)) + b;
         }
     }
 
@@ -48,10 +48,10 @@ public static class EasingFunctions
     /// <param name="c">The amount of change. (end - start).</param>
     /// <param name="d">The total amount of time(milliseconds)/steps.</param>
     /// <returns>The result of the easing function.</returns>
-    public static double EaseInQuad(double t, double b, double c, double d)
+    public static float EaseInQuad(float t, float b, float c, float d)
     {
-        t /= 1000.0; // Convert to seconds
-        d /= 1000.0; // Convert to seconds
+        t /= 1000.0f; // Convert to seconds
+        d /= 1000.0f; // Convert to seconds
 
         t /= d;
